requirements-train.txt
torch>=2.1.0
transformers>=4.36.0
mlflow>=2.10.0
earthengine-api>=0.1.390
geopandas>=0.14.0
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
pydantic>=2.5.0
pyyaml>=6.0
shapely>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
